Revert "[Datasets] Implement push-based shuffle (#23758)"
This reverts commit c1054a0baaea0903a7bcc858fdfa3b5f4f583567.
